requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.9
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.5
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
